添加起送量和单位字段

+++ live_moudle/src/main/java/com/example/live_moudle/live/commodity/selectCommodityProperty/SelectCommodityPropertyDialogFragment.java
@@ -193,7 +193,7 @@ public class SelectCommodityPropertyDialogFragment extends BaseBottomSheetFragme
 
     public interface OnConfirmCallBack {
 
-        void confirm(CommodityPropertyBean commodityPropertyBean, int amount);
+        void confirm(CommodityPropertyBean commodityPropertyBean, double amount);
     }
